+++ modules/apps/layout/layout-content-page-editor-web-test/src/testIntegration/resources/com/liferay/layout/content/page/editor/web/util/dependencies/fragment_composition_with_a_card.json
